Mob #21
Started the refactor from single instance of Player to List<Player>.

Completed moving all Player behavior into Player and cleaning up the public "API" for Player, making methods private as much as possible.

Fixed WebConfigurationTest so that it uses a mock instead of the real HttpGameMonitor so tests don't get stuck waiting for the remote service to wake up.